Moved CardDeck to Components
Moved CardDeck to Component as it is not aware of the Redux state.
Moved Card into CardDeck folder.
