+++ front_end/src/components/CardDeck/Card/Card.module.css
