+++ front_end/src/components/CardDeck/CardDeck.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { configure, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { CardDeck } from './CardDeck';
-import Card from '../../components/Card/Card';
+import Card from './Card/Card';
 import PaginateButtons from
     '../../components/UI/PaginateButtons/PaginateButtons';
